refactor: Refactoring database and file cases for v2 to maintain organization in code
