+++ Database/SQL/Metrics/insert_metric.sql
