+++ Database/SQL/Metrics/select_metrics_by_model_and_domain.sql
