+++ Database/SQL/Structure/create_table_metrics_info.sql
