+++ Database/SQL/Structure/create_table_summary_info.sql
