+++ Database/SQL/Summary/insert_summary.sql
