+++ Database/SQL/Summary/select_all_summaries.sql
