+++ Database/SQL/Summary/select_all_summaries_with_anomaly.sql
