+++ Database/SQL/Summary/select_summary_by_name.sql
